HBASE-28152. Replace scala.util.parsing.json with org.json4s.jackson
In https://issues.apache.org/jira/browse/HBASE-28137 to support Spark 3.4 the scala-parser-combinators was added as direct dependency to HBase Spark Connector.
This was needed as Spark 3.4 is not using scala-parser-combinators and it is not inherited as transitive dependency.

But this solution has a disadvantage. As the HBase Spark Connector assembly jar does not include any 3rd party libraries the scala-parser-combinators must be added to the spark classpath for HBase Spark Connector to work.
A much better solution is to replace scala.util.parsing.json with org.json4s.jackson which is used by Spark core, see https://github.com/apache/spark/blob/branch-3.4/core/pom.xml#L279-L280.